+++ src/tree-graph.ts
@@ -5,6 +5,22 @@ const straightChar = '┃ ';
 const branchAndStraightChar = '┣ ';
 
 /**
+ * Create something like `cli-name/command/another/`
+ * @param treePath The path to the command, including app name
+ * @param colored Should it be colored
+ * @returns Thing like `cli-name/command/another/`
+ */
+function renderTreePath(treePath: string[], colored: boolean): string {
+    treePath.pop(); // remove the last element  because it will be added by the command
+    if (treePath.length === 0) {
+	    return '';
+    }
+    return (colored ? '\x1b[2m' : '') // dimmed
+    + treePath.join('/') + '/'
+    + (colored ? '\x1b[0m' : ''); // reset
+}
+
+/**
  * Generate graph recursively
  * @param command The command to get subcommands from
  * @param margin Count of `┃ ` repeats
@@ -44,7 +60,7 @@ function recursiveTree(command: Command, margin: number, options: TreeGraphOptio
                 // if subcommands are hidden and showing them is not configured
                 if ((subcommand.hideSubcommands ?? false) && !options.showHiddenSubcommands) {
                     // sign that it has hidden subcommands
-		    const sign = options.colored ? ' \x1b[34m+\x1b[0m' : ' +';
+		            const sign = options.colored ? ' \x1b[34m+\x1b[0m' : ' +';
                     result[result.length - 1] += sign;
                 } else {
                     result.push(recursiveTree(subcommand, margin + 1, options, endedMargin));
@@ -62,6 +78,8 @@ export type TreeGraphOptions = {
     showHidden?: boolean,
     /** Show hidden subcommands of commands instead of hiding them under ` +` */
     showHiddenSubcommands?: boolean,
+    /** Add something like `cli-name/command/another/` */
+    addTreePath?: boolean,
 };
 
 /**
@@ -70,15 +88,33 @@ export type TreeGraphOptions = {
  * @options Options for the tree graph generator. Boolean values are legacy, they represent the color
  * @returns The graph
  */
-export function treeGraph(command: Command, commandName: string, options: TreeGraphOptions = {}): string {
+export function treeGraph(command: Command, commandNameOrTreePath: string | string[], options: TreeGraphOptions = {}): string {
     // legacy
     if (typeof options === 'boolean') options = { colored: options };
     // default values
     options.colored ??= false;
     options.showHidden ??= false;
     options.showHiddenSubcommands ??= false;
+    options.addTreePath ??= true;
+    // tree path variables
+    let commandName: string;
+    let treePath: string[];
+    if (typeof commandNameOrTreePath === "string") {
+        commandName = commandNameOrTreePath;
+        treePath = [ commandNameOrTreePath ];
+    } else {
+        commandName = commandNameOrTreePath[commandNameOrTreePath.length - 1];
+        treePath = commandNameOrTreePath;
+    }
+    // tree path render
+    let treePathRender: string;
+    if (options.addTreePath) {
+        treePathRender = renderTreePath(treePath, options.colored);
+    } else {
+        treePathRender = '';
+    }
     // run
     const graph = recursiveTree(command, 0, options, 0);
     const emptyMessage = options.colored ? '\x1b[34m(empty)\x1b[0m' : '(empty)';
-    return commandName + '\n' + (graph || emptyMessage);
+    return treePathRender + commandName + '\n' + (graph || emptyMessage);
 }